requirements.txt
asgiref==3.3.1
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-filter==2.4.0
folium==0.12.0
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
